[IMP] copy links & doc from enterprise wiki, draft
Select links and pieces of information from Odoo Enterprise wiki
were copied over at the end Chap. 2 of the functional training.

I also added some new links, and re-worded part of the text to
better fit the different context (IMHO).

As stated in #1473, I believe these would be useful to have during
the tutorial and not only at the end of it.